+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cat48: Add Mode1CodeOctalRepresentation
 - Cat48: Add Mode2CodeOctalRepresentation
 - Cat48: Add Mode1CodeConfidenceIndicator
+- Cat48: Add Mode2CodeConfidenceIndicator
 - Update Data Item docs
 
 ## [0.2.0] - 2020-08-21